+++ org.eclipse.tracecompass.lttng2.control.ui.tests/stubs/org/eclipse/tracecompass/internal/lttng2/control/stubs/dialogs/CreateSessionDialogStub.java
@@ -32,7 +32,7 @@ public class CreateSessionDialogStub implements ICreateSessionDialog {
     private String fDataUrl = null;
     private boolean fIsSnapshot;
     private boolean fIsLive = false;
-    private int fLiveDelay = 0;
+    private long fLiveDelay = 0;
 
     @Override
     public void initialize(TraceSessionGroup group) {
